==> documents/genomics/parvicapture/04-Baits.Rmd <==
```{r setup_baits, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(parallel)
library(tidyverse)
library(kableExtra)
library(Biostrings)
library(GenomicAlignments)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
path <- "../../../data/Eschweilera_Paracou/Baits/files-Scmitt"
```

# Baits

```{bash, eval=F, echo=T}
cd ../../../data/Eschweilera_Paracou/Baits/files-Scmitt
blat all-targets.masked.fas combined_baits_filtered_19957.fas BaitsonTargets.psl
~/Tools/samtools-1.7/misc/psl2sam.pl BaitsonTargets.psl | samtools view -bt all-targets.masked.fas.fai -o BaitsonTargets.bam 
```

```{r baitsAlignment, fig.cap = "Baits coverage by target type."}
aln <- readGAlignments(file.path(path, "BaitsonTargets.bam"), use.names = T)
lapply(coverage(aln)@listData, function(x) data.frame(length = x@lengths, coverage = x@values)) %>% 
  bind_rows(.id = "seq") %>% 
  mutate(type = ifelse(grepl("FT", seq), "FT", "NT")) %>% 
  group_by(seq, coverage) %>% 
  do(data.frame(seq = rep(.$seq, .$length), 
                type = rep(.$type, .$length),
                coverage = rep(.$coverage, .$length), 
                stringsAsFactors = FALSE)) %>% 
  ggplot(aes(coverage, fill = type)) + 
  geom_histogram(position = "dodge") + scale_y_log10()
```
