==> documents/genomics/parvicapture/08-Capture.Rmd <==
```{r setup_capture, include=FALSE}
rm(list = ls()) ; invisible(gc()) ; set.seed(42)
library(knitr)
library(kableExtra)
if(knitr:::is_html_output()) options(knitr.table.format = "html") 
if(knitr:::is_latex_output()) options(knitr.table.format = "latex") 
library(parallel)
library(tidyverse)
library(rstan)
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = T)
theme_set(bayesplot::theme_default())
opts_chunk$set(
  echo = F, message = F, warning = F, fig.height = 6, fig.width = 8,
  cache = T, cache.lazy = F)
path <- "../../../data/Eschweilera_Paracou"
```

# Capture

After library preparation, we prepared plates for capture. We proceed as follow :

1. __Dosage__: Obtained libraries were dosed by Qubit after range assessment by NanoDrop (16 libraries from P5)
1. __Repeats__: X libraries were repeated from extraction to replace least concentrated libraries
1. __Equimolar pools preparation__: Equimolar pools were built for capture
1. __Pools fragment size evaluation__: Pool fragents sizes were evalutaed by Bioanalyzer
1. __Pools fragment size selection__: Pool fragments size were filtered by Pippin (failed) and then by beads (due to pippin failure) between 200 and 700 bp, but with a narrow mode around 350 bp.
1. __Capture__: ...

## Dosage

Obtained libraries were dosed by Qubit after range assessment by NanoDrop (16 libraries from P5). Most of libraries had a concentration above 10 $ng.\mu L^{-1}$ and libraries had consistent degradation pattern with electrophoresis.

```{r quantit, fig.cap="Library concentration measured by QuantIt."}
quantit1 <- read_csv(file.path(path, "QuantIt", 
                   "PARVICAPTURE_library_P1P2P5_02072019_fluo.csv"), skip = 53)[2:4] %>% 
  dplyr::rename(position = Well, fluo = `dosage_fluo_384_CB:501,527`) %>% 
  dplyr::select(position, fluo) %>% 
  filter(!is.na(fluo)) %>% 
  left_join(read_csv(file.path(path, "QuantIt", 
                               "PARVICAPTURE_library_P1P2P5_02072019_plates.csv"),  skip = 36) %>% 
              dplyr::rename(row = X1) %>% 
              reshape2::melt(id.vars = "row", variable.name = "col", value.name = "library") %>% 
              mutate(position = paste0(row, col)) %>% 
              select(position, library)) %>% 
  separate(library, c("Plate_library", "Position_library"), "_") %>% 
  mutate(Plate_library = as.numeric(gsub("P", "", Plate_library))) %>% 
  select(Plate_library, Position_library, fluo) %>% 
  arrange(Plate_library, Position_library) %>% 
  left_join(data.frame(Plate_library = 0,
                       Position_library = as.vector(sapply(1:6, function(x) paste0(LETTERS[1:4], x))),
                       concentration = rep(c(0, 10, 20, 40, 80, 120), each = 4))) %>% 
  mutate(QuantIt = 1)
data1 <- filter(quantit1, Plate_library == 0)
quantit1 <- quantit1 %>% 
  mutate(concentration_est = lm(concentration ~ fluo, data = data1)$coefficients[2]*fluo)
quantit2 <- read_csv(file.path(path, "QuantIt", 
                   "PARVICAPTURE_libraryP3P4_20190826_fluo.csv"), skip = 53)[2:4] %>% 
  dplyr::rename(position = Well, fluo = `dosage_fluo_384_CB:501,527`) %>% 
  dplyr::select(position, fluo) %>% 
  filter(!is.na(fluo)) %>% 
  left_join(read_csv(file.path(path, "QuantIt", 
                               "PARVICAPTURE_libraryP3P4_20190826_plates.csv"),  skip = 36) %>% 
              dplyr::rename(row = X1) %>% 
              reshape2::melt(id.vars = "row", variable.name = "col", value.name = "library") %>% 
              mutate(position = paste0(row, col)) %>% 
              select(position, library)) %>% 
  separate(library, c("Plate_library", "Position_library"), "_") %>% 
  mutate(Plate_library = as.numeric(gsub("P", "", Plate_library))) %>% 
  select(Plate_library, Position_library, fluo) %>% 
  arrange(Plate_library, Position_library) %>% 
  left_join(data.frame(Plate_library = 0,
                       Position_library = as.vector(sapply(1:6, function(x) paste0(LETTERS[1:4], x))),
                       concentration = rep(c(0, 10, 20, 40, 80, 120), each = 4))) %>% 
  mutate(QuantIt = 2)
data2 <- filter(quantit2, Plate_library == 0)
quantit2 <- quantit2 %>% 
  mutate(concentration_est = lm(concentration ~ fluo, data = data2)$coefficients[2]*fluo)
quantit <- bind_rows(quantit1, quantit2) %>% 
  filter(Plate_library != 0)
data <- bind_rows(data1, data2)
rm(quantit1, quantit2, data1, data2)
ggplot(quantit, aes(concentration_est, fill = concentration_est > 10)) +
  geom_histogram() +
  scale_x_log10() +
  xlab("Concentration") +
  ylim(0, 45) +
  theme(legend.position = "bottom") +
  annotation_custom(grob = ggplotGrob(ggplot(data, aes(fluo, concentration,
                                                       col = QuantIt, group = QuantIt)) +
                                        geom_point() +
                                        geom_smooth(method = "lm") +
                                        xlab("Fluorescence") +
                                        ylab("Concentration") +
                                        scale_color_continuous(guide = "none")), 
                    xmin = log10(10), xmax = log10(100), ymin = 28, ymax = 45)
```

## Repeats & Reamplifcations

209 libraries have an estimated concentration below 10 $ng.\mu L ^{-1}$, but we have only 96 spot on plate 6 corresponding to a minimal estimated concentration of 4.4 $ng.\mu L ^{-1}$. In addition to repeats, library between an estimated concentration of 4.4 and 8 $ng.\mu L ^{-1}$ have been reamplified with 8 cycles. Repeats have more or less worked with medium concentration of DNA whereas reamplification gave good results (Fig. \@ref(fig:quantit2)).

```{r repeats2, eval=F, echo=F}
quantit %>% 
  left_join(googlesheets::gs_title("Parvicapture") %>% 
             googlesheets::gs_read("Extraction"),
            by = c("Plate_library" = "Plate_extraction",
                   "Position_library" = "Position_extraction")) %>% 
  filter(concentration_est < 4.4) %>% 
  arrange(Plate_library, Position_library) %>% 
  mutate(Plate_repeat = 6) %>% 
  mutate(Position_repeat = as.vector(sapply(1:12, function(x) paste0(LETTERS[1:8], x)))) %>% 
  mutate(Repeat = 1) %>% 
  select(IdGenetic, Plate_library, Position_library, Plate_repeat, Position_repeat, Genus, Species, Plot, SubPlot, TreeFieldNum, Repeat) %>% 
  write_tsv(path = file.path(path, "repeat.tsv"))
```

```{r repeatNano, eval=F}
extraction <- googlesheets::gs_title("Parvicapture") %>% 
  googlesheets::gs_read("Extraction")
read.delim2(file.path(path, "NanoDrop", "Purification", "PARVICAPTURE_library_P6_20190917.txt")) %>% 
  mutate(Plate_extraction = 6) %>% 
  mutate(Conc. = ifelse(Conc. < 0, 0, Conc.)) %>% 
  dplyr::rename(Position_extraction = Well,
                Concentration_Purified = Conc.) %>% 
  mutate(Plate_extraction = as.numeric(Plate_extraction)) %>% 
  select(Plate_extraction, Position_extraction, Concentration_Purified) %>% 
  ggplot(aes(48+0.15*Concentration_Purified)) +
  geom_histogram() +
  scale_x_log10()
```

```{r reamplification, eval=F, echo=F}
quantit %>% 
  left_join(googlesheets::gs_title("Parvicapture") %>% 
             googlesheets::gs_read("Extraction"),
            by = c("Plate_library" = "Plate_extraction",
                   "Position_library" = "Position_extraction")) %>% 
  filter(concentration_est > 4.4, concentration_est < 8) %>% 
  arrange(Plate_library, Position_library) %>% 
  mutate(Plate_reamplification = 7) %>% 
  mutate(row = substr(Position_library, 1, 1)) %>% 
  mutate(col = as.numeric(substr(Position_library, 2, 3))) %>% 
  arrange(Plate_library, col, row) %>% 
  mutate(Position_reamplification = as.vector(sapply(1:12, function(x) paste0(LETTERS[1:8], x)))) %>% 
  mutate(Repeat = 1) %>% 
  select(IdGenetic, Plate_library, Position_library, Plate_reamplification, Position_reamplification, Genus, Species, Plot, SubPlot, TreeFieldNum, Repeat) %>% 
  write_tsv(path = file.path(path, "reamplification.tsv"))
```

```{r quantit2, fig.cap="Library repeats and reamplification concentration measured by QuantIt."}
quantit3 <- read_csv(file.path(path, "QuantIt", 
                   "PARVICAPTURE_library_P6P7_20190919_fluo.csv"), skip = 53)[2:4] %>% 
  dplyr::rename(position = Well, fluo = `dosage_fluo_384_CB:501,527`) %>% 
  dplyr::select(position, fluo) %>% 
  filter(!is.na(fluo)) %>% 
  left_join(read_csv(file.path(path, "QuantIt", 
                               "PARVICAPTURE_library_P6P7_20190919_plates.csv"),  skip = 36) %>% 
              dplyr::rename(row = X1) %>% 
              reshape2::melt(id.vars = "row", variable.name = "col", value.name = "library") %>% 
              mutate(position = paste0(row, col)) %>% 
              select(position, library)) %>% 
  separate(library, c("Plate_library", "Position_library"), "_") %>% 
  mutate(Plate_library = as.numeric(gsub("P", "", Plate_library))) %>% 
  select(Plate_library, Position_library, fluo) %>% 
  arrange(Plate_library, Position_library) %>% 
  left_join(data.frame(Plate_library = 0,
                       Position_library = as.vector(sapply(1:6, function(x) paste0(LETTERS[1:4], x))),
                       concentration = rep(c(0, 10, 20, 40, 80, 120)*21/20, each = 4))) %>% 
  mutate(QuantIt = 3)
data3 <- filter(quantit3, Plate_library == 0)
quantit3 <- quantit3 %>% 
  mutate(concentration_est = lm(concentration ~ fluo, data = data3)$coefficients[2]*fluo)
quantit <- bind_rows(quantit, quantit3) %>% 
  filter(Plate_library != 0)
data <- bind_rows(data, data3)
ggplot(filter(quantit3, Plate_library != 0), aes(concentration_est, fill = as.factor(Plate_library))) +
  geom_histogram() +
  scale_x_log10() +
  xlab("Concentration") +
  ylim(0, 45) +
  theme(legend.position = "bottom") +
  annotation_custom(grob = ggplotGrob(ggplot(data3, aes(fluo, concentration,
                                                       col = QuantIt, group = QuantIt)) +
                                        geom_point() +
                                        geom_smooth(method = "lm") +
                                        xlab("Fluorescence") +
                                        ylab("Concentration") +
                                        scale_color_continuous(guide = "none")), 
                    xmin = log10(10), xmax = log10(100), ymin = 28, ymax = 45)
```

## Equimolar pools preparation

All samples have been pulled by batch of 134 following plate order for the 4 reactions of Pippin (with 1 repeat). We wanted 100 to 500 $ng$ of DNA per capture reaction, we thus  used $8 ng.\mu L^{-1} \times 20 \mu L = 160 ng$ of DNA of each sample ($8 ng.\mu L^{-1}$ being in the lower concentrations, samples below will be sacrified and considered as 8 because they are repeated), resulting in $160 ng \times 128 = 20.48 \mu g$ of DNA per pool (because we will loose material in purification, we should aim for extra).

```{r pools}
rm(quantit3, data3)
quantit <- quantit %>% 
  dplyr::select(-concentration, -fluo, -QuantIt) %>% 
  dplyr::rename(concentration = concentration_est)
extraction <- googlesheets::gs_title("Parvicapture") %>% 
             googlesheets::gs_read("Extraction")
reamplification <- googlesheets::gs_title("Parvicapture") %>% 
             googlesheets::gs_read("Reamplification")
data <- extraction %>% 
  left_join(quantit, by = c("Plate_extraction" = "Plate_library",
                            "Position_extraction" = "Position_library")) %>% 
  left_join(reamplification %>% 
              left_join(quantit, by = c("plate_reamplifacation" = "Plate_library",
                                        "Position_reamplifcation" = "Position_library")) %>% 
              mutate(concentration_reamplification = concentration) %>% 
              dplyr::select(Plate_library, Position_library, concentration_reamplification),
            by = c("Plate_extraction" = "Plate_library",
                            "Position_extraction" = "Position_library")) %>% 
  mutate(concentration = ifelse(is.na(concentration_reamplification), concentration, concentration_reamplification)) %>% 
  dplyr::select(-concentration_reamplification) %>% 
  filter(!(Plate_extraction == 5 & Repeat == 1)) %>% 
  mutate(Pool_Pippin = c(rep(LETTERS[1:3], each = ceiling(4*96/3))[1:(4*96)], rep("D", 534-4*96)))%>% 
  # mutate(Pool_Pippin = rep(LETTERS[1:4], each = ceiling(nrow(.)/4))[1:nrow(.)])%>% 
  mutate(dilution = 1) %>% 
  mutate(volume = round(ifelse(concentration < 8, 20,
                         160/concentration),1))
rm(extraction, reamplification, quantit) 
```

```{r poolsPlate, fig.cap="Sample pools tube per plate."}
data %>%
  mutate(row = substr(Position_extraction, 1, 1)) %>% 
  mutate(col = substr(Position_extraction, 2, 3)) %>% 
  ggplot(aes(y = factor(row,levels = rev(LETTERS[1:8])),
             x = factor(col,levels = 1:12))) + 
  geom_point(aes(col = Pool_Pippin), size =6)  +
  geom_text(aes(label = Pool_Pippin), size = 2.5, col = "white") +
  facet_wrap(~ Plate_extraction, ncol = 2, labeller = "label_both") +
  theme_bw() +
  labs(x=NULL, y = NULL) 
```

```{r poolsVolume, fig.cap="Sample reaction volume per plate."}
data %>%
  mutate(row = substr(Position_extraction, 1, 1)) %>% 
  mutate(col = substr(Position_extraction, 2, 3)) %>% 
  ggplot(aes(y = factor(row,levels = rev(LETTERS[1:8])),
             x = factor(col,levels = 1:12))) + 
  geom_point(aes(col = volume), size =6)  +
  geom_text(aes(label = volume), size = 2.5, col = "white") +
  facet_wrap(~ Plate_extraction, ncol = 2, labeller = "label_both") +
  theme_bw() +
  labs(x=NULL, y = NULL) 
```

```{r poolsTab, eval=F}
data %>% 
  select(Pool_Pippin, Plate_extraction, Position_extraction, volume) %>% 
  kable(caption = "Capture reactions.",
        col.names = c("Pool", "Plate", "Position","Volume"),
        format = "pandoc", escape = F) %>% 
  kable_styling(full_width = F) 
```

## Pools fragment size evaluation

Fragments size distribution was almost homogeneous between pools, only pool **D** included a fifth of it's DNA in a pike around 150 bp (Fig. \@ref(fig:TapeStation)). We used Pippin and beads to select fragments between 250 and 700 bp.

```{r TapeStation, fig.cap="Pools size assessment pre Pippin."}
knitr::include_graphics("parvicapture_data/TapeStationPrePippin.png")
```

Pool fragents sizes were evalutaed by Bioanalyzer

## Pools fragment size selection

### Pippin

Pool fragments size were filtered by Pippin between 250 and 700 bp. Pippin encountered a bug and didn't detect the ladder pike at 600 bp. Consequently the elution lasted too long. Nevertheless it does not explain the odd pattern observed with TapeStation (Fig. \@ref(fig:TapeStationPostPippin)). We almost completly lost the pool **D** and all other pools have been cut around 450 bp which is too high. Several things might had a role: DNA was high (4 $\mu g$ with upper limit at 5 $\mu g$), but it should have result in a lower cut not higher; and during concentration of pools from up to 800 $\mu L$ to 30 $\mu L$ we also concentrated TE buffer salts. Anyway the clear cut indicated clearly that the size selection resulted from pippin and not cleaning beads. Consequently, we considered this pippin as a failure and decided to develop a protocol with beads to size select pools with the pools originally scheduled as pippin repeat.

```{r TapeStationPostPippin, fig.cap="Pools size assessment post Pippin."}
knitr::include_graphics("parvicapture_data/TapeStationPostPippin1.png")
```

### Beads test

We tested size selection with AMPure beads. The idea is to (i) capture long fragments (> 700 bp) with a small volume of beads,  (ii) transfer supernatant without long fragments, and (iii) capture intermediate fragments (> 250 bp) with beads and discard supernatant. The complicated step is the selection of the second beads volume, as supernatant already included PEG and salts depending on the first beads volume. Consequently, we first tested different first beads volume from 0.55X to 0.4X (`r paste(round(seq(0.55, 0.4, length.out = 8),2), collapse = " ")`) with a second beads volume of 1X, resulting to 0.53X cuting below 800bp (Fig. \@ref(fig:lowerCut)). Finally, we tested different second beads volume from 0.30X to 0.20X (`r paste(round(seq(0.3, 0.2, length.out = 8),2), collapse = " ")`) with a first beads volume of 0.53X, resulting to 0.3X cutting above 300bp (Fig. \@ref(fig:upperCut)). As results were not concluent, we did a third test with a first cut between 0.53, 0.55, 0.58, and 0.6 and a second cut between 0.4 and 0.5 resulting in 8 test samples. Size selection with a first cut at 0.6X and a second at 0.4X seemed to have the expected result with libraries between 250 and 650 bp (Fig. \@ref(fig:finalCut)). We made the test with 4 $\mu L$ of 1kb ladder in 96 $\mu L$ of $H_2OmQ$.

```{r lowerCut, fig.cap="Beads size selection test I."}
knitr::include_graphics("parvicapture_data/ElectroTestI.png")
```

```{r upperCut, fig.cap="Beads size selection test II.."}
knitr::include_graphics("parvicapture_data/ElectroTestII.png")
```

```{r finalCut, fig.cap="Beads size selection test III."}
knitr::include_graphics("parvicapture_data/ElectroTestIII.png")
```

## Beads selection

Beads selection have been realized on splitted pools (originally Pippin2), while keeping every step material in case we need to recover material. We removed long fragment > 700 bp with 0.6X AMPure and selected fragment above 200 bp with 0.4X AMPure. Fragments size distribution was almost homogeneous around 350 bp, except for a very small quantity of small DNA in pool D around 151 bp (Fig. \@ref(fig:TapeStationPostBeads)). We obtained between 1536 and 688 $ng$ of DNA, enough to proceed to 4 repeats of capture per pool with 172 to 384 $ng$ of DNA (Tab. \@ref(tab:SizeSelectionTab)).

Size selection with beads is a bit more permissive than Pippin, that's why we obtained a wider size distribution than expected and still have small distribution around 151 bp for pool D. Consequently, using beads we have to aim for a narrower spectrum and we thus lose more material (mean loss coeficient of 4, Tab. \@ref(tab:SizeSelectionTab)). On the other hand, size selection with beads resulted in a Gaussian distribution with a smaller variance with more fragments around 350 bp, which may be better for capture and sequencing.

```{r TapeStationPostBeads, fig.cap="Pools size assessment post beads."}
knitr::include_graphics("parvicapture_data/TapeStationPostBeads.png")
```

```{r SizeSelectionTab}
data.frame(Pool = LETTERS[1:4], No = 4000, Qubit = c(3.84, 2.58, 1.72, 2.31)) %>% 
  mutate(Nf = Qubit*10*40) %>% 
  mutate(LossCoef = round(No/Nf)) %>% 
  mutate(NCapture = Nf/4) %>% 
  dplyr::select(-Qubit) %>% 
  kable(col.names = c("Pool", "N origin (µg)", "N final (µg)", "Loss coefficient", "N capture (µg)"),
        caption = "Beads size selection results.")
```

## Capture

We split the 4 pools in 16 reactions (4 replicates for each) and realized the capture following [ArborScience protocol](https://arborbiosci.com/wp-content/uploads/2018/04/myBaits-Manual-v4.pdf). We obtained between 5 and 37 ng of DNA per pool assessed by Qubit (with the four replicates pooled). We thus built the final pools for sequencing lanes. **Need to resolve pool _D_ lack of DNA.**  We thus obtained XX $\mu L$ of Lane 1 at $XX~nM$ and XX $\mu L$ of Lane 2 at $XX~nM$ ([see google sheets for more details](https://docs.google.com/spreadsheets/d/1nTx_mofzpTrnuv9qQY8y99RTxsdhLfDyRy2k3s0Tk4A/edit?usp=sharing)). The material has been sent to IGA for sequencing on Illumina HiSeq 3000 on two lanes of pair-ends 150 bp sequences.
